services/runner/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nexus-runner"
version = "0.1.0"
description = "Nexus SaaS worker runner"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "python-jose[cryptography]>=3.4.0",
  "pydantic>=2.10.6",
  "pydantic-settings>=2.7.1",
  "redis>=6.0.0",
  "websockets>=15.0",
]

[project.optional-dependencies]
test = [
  "pytest>=8.3.4",
  "pytest-asyncio>=0.25.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
